Create dbus_consts.rs and put DBus consts in there
This is most of what was in consts.rs.

Also, use allow(non_camel_case_types) to silence warnings in it.
